feat: CFD-306 Netex updates for prices per SOP (#55)
