+++ repos/fdbt-netex-output/.eslintrc.js
@@ -7,7 +7,6 @@ module.exports = {
     extends: [
         'plugin:@typescript-eslint/recommended',
         'prettier/@typescript-eslint',
-        'airbnb-base',
         'plugin:prettier/recommended',
         'plugin:@typescript-eslint/recommended-requiring-type-checking',
         'eslint-config-prettier',
@@ -34,15 +33,6 @@ module.exports = {
                 vars: 'all',
                 args: 'after-used',
                 ignoreRestSiblings: false,
-            },
-        ],
-        'no-console': 'warn',
-        'import/extensions': [
-            'error',
-            'ignorePackages',
-            {
-                js: 'never',
-                ts: 'never',
             },
         ],
         'no-console': 0,
