+++ repos/fdbt-netex-output/src/netex-convertor/handler.ts
@@ -120,6 +120,8 @@ export const netexConvertorHandler = async (event: S3Event): Promise<void> => {
             console.info(`NeTEx generation complete for type ${scheme}${carnet}${type}`);
         }
     } catch (error) {
+        console.error(error);
+
         const sns: SNS = new AWS.SNS();
 
         const messageParams = {
@@ -134,7 +136,9 @@ export const netexConvertorHandler = async (event: S3Event): Promise<void> => {
             },
         };
 
-        await sns.publish(messageParams).promise();
+        if (process.env.STAGE !== 'dev') {
+            await sns.publish(messageParams).promise();
+        }
 
         throw error;
     }
