+++ repos/fdbt-netex-output/src/netex-convertor/run-local.ts
@@ -2,6 +2,9 @@ import netexConvertorHandler from './handler';
 
 const event = JSON.parse(process.argv.slice(2)[0]);
 
+process.env.SNS_ALERTS_ARN = 'arn:aws:sns:us-east-1:000000000000:AlertsTopic';
+process.env.STAGE = 'dev';
+
 netexConvertorHandler(event)
     .then(() => process.exit(0))
     .catch(error => {
