+++ repos/fdbt-netex-output/src/types/index.ts
@@ -1,3 +1,4 @@
+import { SalesOfferPackage } from '../../../../shared/matchingJsonTypes';
 import { NetexObject } from '../netex-convertor/sharedHelpers';
 
 // Misc
@@ -79,14 +80,6 @@ export type Ticket = SpecificTicket &
         groupDefinition: { maxPeople?: string; companions?: GroupCompanion[] };
         carnet: boolean;
     }>;
-
-export interface SalesOfferPackage {
-    name: string;
-    description: string;
-    purchaseLocations: string[];
-    paymentMethods: string[];
-    ticketFormats: string[];
-}
 
 export interface TimeBand {
     startTime: string;
