+++ repos/fdbt-site/src/interfaces/index.ts
@@ -2,6 +2,7 @@ import { NextApiRequest, NextPageContext } from 'next';
 import { DocumentContext } from 'next/document';
 import { IncomingMessage, ServerResponse } from 'http';
 import { ReactElement } from 'react';
+import { SalesOfferPackage } from '../../../../shared/matchingJsonTypes';
 
 // Session Attributes and Cookies
 
@@ -537,12 +538,6 @@ export interface SalesOfferPackageInfo {
 
 export interface SalesOfferPackageInfoWithErrors extends SalesOfferPackageInfo {
     errors: ErrorInfo[];
-}
-
-export interface SalesOfferPackage extends SalesOfferPackageInfo {
-    name: string;
-    description: string;
-    id?: string;
 }
 
 export interface SalesOfferPackageWithErrors extends SalesOfferPackage {
