+++ repos/fdbt-site/src/pages/selectSalesOfferPackage.tsx
@@ -1,4 +1,5 @@
 import React, { ChangeEventHandler, ReactElement, useState } from 'react';
+import { SalesOfferPackage } from '../../../../shared/matchingJsonTypes';
 import CsrfForm from '../components/CsrfForm';
 import DeleteSOPButton from '../components/DeleteSOPButton';
 import ErrorSummary from '../components/ErrorSummary';
@@ -15,7 +16,6 @@ import {
     NextPageContextWithSession,
     ProductInfo,
     ProductWithSalesOfferPackages,
-    SalesOfferPackage,
     SchoolFareTypeAttribute,
 } from '../interfaces';
 import { isFareType } from '../interfaces/typeGuards';
