+++ shared/matchingJsonTypes.d.ts
@@ -1,0 +1,9 @@
+export interface SalesOfferPackage {
+  id?: string;
+  name: string;
+  description: string;
+  purchaseLocations: string[];
+  paymentMethods: string[];
+  ticketFormats: string[];
+  price?: string;
+}
