+++ shared/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../repos/fdbt-site",
+  "include": ["./**"]
+}
